+++ src/septentrio_gnss_driver/communication/rx_message.cpp
@@ -30,6 +30,8 @@
 
 #include <thread>
 
+#include<boost/tokenizer.hpp>
+
 #include <GeographicLib/UTMUPS.hpp>
 
 #include <septentrio_gnss_driver/communication/rx_message.hpp>
@@ -2607,6 +2609,68 @@ bool io_comm_rx::RxMessage::read(std::string message_key, bool search)
 				node_->log(LogLevel::ERROR, "septentrio_gnss_driver: parse error in ReceiverSetup");
 				break;
 			}
+			static int32_t ins_major  = 1;
+			static int32_t ins_minor  = 3;
+			static int32_t ins_patch  = 2;
+			static int32_t gnss_major = 4;
+			static int32_t gnss_minor = 10;
+			static int32_t gnss_patch = 0;
+			boost::tokenizer<> tok(last_receiversetup_.rx_version);
+			boost::tokenizer<>::iterator it = tok.begin();
+			std::vector<int32_t> major_minor_patch;
+			major_minor_patch.reserve(3);
+			for (boost::tokenizer<>::iterator it = tok.begin(); it!=tok.end(); ++it)
+			{
+				int32_t v = std::atoi(it->c_str());
+				major_minor_patch.push_back(v);
+			}
+			if (major_minor_patch.size() < 3)
+			{
+				node_->log(LogLevel::ERROR, "septentrio_gnss_driver: parse error of firmware version.");
+			}
+			else
+			{
+				if (settings_->septentrio_receiver_type == "ins")
+				{
+					if ((major_minor_patch[0] < ins_major) ||
+						((major_minor_patch[0] == ins_major) && (major_minor_patch[1] < ins_minor)) ||
+						((major_minor_patch[0] == ins_major) && (major_minor_patch[1] == ins_minor) && (major_minor_patch[2] < ins_patch)))
+					{
+						node_->log(LogLevel::WARN, "INS receiver has firmware version: " + last_receiversetup_.rx_version + 
+												   ", which does not support all features. Please update to at least " + 
+												   std::to_string(ins_major) + "." + std::to_string(ins_minor) + "." + 
+												   std::to_string(ins_patch) + " or consult README.");
+					}
+				}
+				if (settings_->septentrio_receiver_type == "gnss")
+				{
+					if (major_minor_patch[0] < 3)
+					{
+						// Assuming INS is used as GNSS
+						if ((major_minor_patch[1] < ins_minor) ||
+							(major_minor_patch[1] == ins_minor) && (major_minor_patch[2] < ins_patch))
+						{
+							node_->log(LogLevel::WARN, "INS receiver has firmware version: " + last_receiversetup_.rx_version + 
+													   ", which does not support all features. Please update to at least " + 
+													   std::to_string(ins_major) + "." + std::to_string(ins_minor) + "." + 
+													   std::to_string(ins_patch) + " or consult README.");							 
+						}
+						node_->log(LogLevel::WARN, "INS receiver seems to be used as GNSS. Some settings may trigger warnings.");
+					}
+					else if ((major_minor_patch[0] < gnss_major) ||
+							((major_minor_patch[0] == gnss_major) && (major_minor_patch[1] < gnss_minor)) ||
+							((major_minor_patch[0] == gnss_major) && (major_minor_patch[1] == gnss_minor) && (major_minor_patch[2] < gnss_patch)))
+					{
+						node_->log(LogLevel::WARN, "GNSS receiver has firmware version: " + last_receiversetup_.rx_version + 
+												   ", which may not support all features. Please update to at least " + 
+												   std::to_string(gnss_major) + "." + std::to_string(gnss_minor) + "." + 
+												   std::to_string(gnss_patch) + " or consult README.");
+					}
+					else
+						node_->log(LogLevel::ERROR, "gnss");
+				}
+			}			
+
 			break;
 		}
 		case evReceiverTime: